move grpc helper function to internal directory
