+++ services/slam/slam.go
@@ -8,7 +8,6 @@ import (
 	"sync"
 
 	"github.com/edaniels/golog"
-	"github.com/pkg/errors"
 	pb "go.viam.com/api/service/slam/v1"
 	goutils "go.viam.com/utils"
 	"go.viam.com/utils/rpc"
@@ -81,26 +80,6 @@ type Service interface {
 	GetPointCloudMapStream(ctx context.Context, name string) (func() ([]byte, error), error)
 	GetInternalStateStream(ctx context.Context, name string) (func() ([]byte, error), error)
 	resource.Generic
-}
-
-// HelperGetInternalStateCallback helps a client request the internal state stream from a SLAM server.
-func HelperGetInternalStateCallback(ctx context.Context, name string, slamClient pb.SLAMServiceClient) (func() ([]byte, error), error) {
-	req := &pb.GetInternalStateStreamRequest{Name: name}
-
-	resp, err := slamClient.GetInternalStateStream(ctx, req)
-	if err != nil {
-		return nil, errors.Wrap(err, "error getting the internal state from the SLAM client")
-	}
-
-	f := func() ([]byte, error) {
-		chunk, err := resp.Recv()
-		if err != nil {
-			return nil, err
-		}
-
-		return chunk.GetInternalStateChunk(), nil
-	}
-	return f, err
 }
 
 type reconfigurableSlam struct {
